change: move `leaderboard.ts` file
